feat(back) CorreÃÃo de problemas nas regras de negÃcios re #49

+++ backend/src/main/java/com/api/caramelo/services/PetService.java
@@ -63,6 +63,10 @@ public class PetService implements IPetService {
             throw new BusinessRuleException("Usuário com esse token não existe.");
         }
 
+        if (userId != petRepository.findPetById(petId).getUser().getId()) {
+            throw new BusinessRuleException("Um usuário não pode atualizar um Pet de outro usuário");
+        }
+
         Optional<Pet> optionalPet = petRepository.findById(petId);
 
         if (optionalPet.isEmpty()) {
@@ -98,7 +102,12 @@ public class PetService implements IPetService {
             throw new BusinessRuleException("Usuário com esse token não existe.");
         }
 
+        if (userId != petRepository.findPetById(petId).getUser().getId()) {
+            throw new BusinessRuleException("Um usuário não pode deletar um Pet de outro usuário");
+        }
+
         petRepository.deleteById(petId);
+
     }
 
     @Override
